refactor: fixes project structure
