+++ projetomultithread/src/animation/animation_config.h
@@ -1,0 +1,11 @@
+#ifndef CONFIG_H
+#define CONFIG_H
+
+#define WIDTH 160
+#define HEIGHT 40
+
+#define TRAIN_MOVE_FPS 120
+
+#define SLEEP_AFTER 2000000
+#define SLEEP_BOARD 500000
+#endif
