+++ projetomultithread/src/rollercoaster.c
@@ -15,150 +15,11 @@
 #include <unistd.h>
 #include <string.h>
 
+#include "sprites/sprites.h"
+#include "animation/animation.h"
+#include "animation/animation_config.h"
 #include "buffer/buffer.h"
 #include "primitives/primitives.h"
-
-#define N_PASSENGER 15
-#define CAPACITY 5
-#define N_RUNS 2
-
-#define WIDTH 160
-#define HEIGHT 40
-
-#define CART_STARTX 10
-#define CART_STARTY 04
-#define CART_PADDING 12
-
-#define QUEUE_STARTX 20
-#define QUEUE_STARTY 29
-
-#define TRAIN_MOVE_FPS 120
-
-#define SLEEP_AFTER 2000000
-#define TRAIN_CD 1000000 / TRAIN_MOVE_FPS
-#define SLEEP_BOARD 500000
-
-// Title Screens
-char *OPENING =
-    " _______  _______  _        _        _______  _______    _______  _______  _______  _______ _________ _______  _______ \n"
-    "(  ____ )(  ___  )( \\      ( \\      (  ____ \\(  ____ )  (  ____ \\(  ___  )(  ___  )(  ____ \\__   __/(  ____ \\(  ____ )\n"
-    "| (    )|| (   ) || (      | (      | (    \\/| (    )|  | (    \\/| (   ) || (   ) || (    \\/   ) (   | (    \\/| (    )|\n"
-    "| (____)|| |   | || |      | |      | (__    | (____)|  | |      | |   | || (___) || (_____    | |   | (__    | (____)|\n"
-    "|     __)| |   | || |      | |      |  __)   |     __)  | |      | |   | ||  ___  |(_____  )   | |   |  __)   |     __)\n"
-    "| (\\ (   | |   | || |      | |      | (      | (\\ (     | |      | |   | || (   ) |      ) |   | |   | (      | (\\ (   \n"
-    "| ) \\ \\__| (___) || (____/\\| (____/\\| (____/\\| ) \\ \\__  | (____/\\| (___) || )   ( |/\\____) |   | |   | (____/\\| ) \\ \\__\n"
-    "|/   \\__/(_______)(_______/(_______/(_______/|/   \\__/  (_______/(_______)|/     \\|\\_______)   )_(   (_______/|/   \\__/\n"
-    "                             _________ _______    _______  _______  _______  _                                         \n"
-    "                             \\__   __/(  ____ \\  (  ___  )(  ____ )(  ____ \\( (    /|                               \n"
-    "                                ) (   | (    \\/  | (   ) || (    )|| (    \\/|  \\  ( |                               \n"
-    "                                | |   | (_____   | |   | || (____)|| (__    |   \\ | |                                 \n"
-    "                                | |   (_____  )  | |   | ||  _____)|  __)   | (\\ \\) |                                \n"
-    "                                | |         ) |  | |   | || (      | (      | | \\   |                                 \n"
-    "                             ___) (___/\\____) |  | (___) || )      | (____/\\| )  \\  |                               \n"
-    "                             \\_______/\\_______)  (_______)|/       (_______/|/    )_)                                \n";
-
-char *ENDING =
-    " _______  _______  _        _        _______  _______    _______  _______  _______  _______ _________ _______  _______ \n"
-    "(  ____ )(  ___  )( \\      ( \\      (  ____ \\(  ____ )  (  ____ \\(  ___  )(  ___  )(  ____ \\__   __/(  ____ \\(  ____ )\n"
-    "| (    )|| (   ) || (      | (      | (    \\/| (    )|  | (    \\/| (   ) || (   ) || (    \\/   ) (   | (    \\/| (    )|\n"
-    "| (____)|| |   | || |      | |      | (__    | (____)|  | |      | |   | || (___) || (_____    | |   | (__    | (____)|\n"
-    "|     __)| |   | || |      | |      |  __)   |     __)  | |      | |   | ||  ___  |(_____  )   | |   |  __)   |     __)\n"
-    "| (\\ (   | |   | || |      | |      | (      | (\\ (     | |      | |   | || (   ) |      ) |   | |   | (      | (\\ (   \n"
-    "| ) \\ \\__| (___) || (____/\\| (____/\\| (____/\\| ) \\ \\__  | (____/\\| (___) || )   ( |/\\____) |   | |   | (____/\\| ) \\ \\__\n"
-    "|/   \\__/(_______)(_______/(_______/(_______/|/   \\__/  (_______/(_______)|/     \\|\\_______)   )_(   (_______/|/   \\__/\n"
-    "                          _________ _______    _______  _        _______  _______  _______  ______                     \n"
-    "                          \\__   __/(  ____ \\  (  ____ \\( \\      (  ___  )(  ____ \\(  ____ \\(  __  \\                    \n"
-    "                             ) (   | (    \\/  | (    \\/| (      | (   ) || (    \\/| (    \\/| (  \\  )                   \n"
-    "                             | |   | (_____   | |      | |      | |   | || (_____ | (__    | |   ) |                   \n"
-    "                             | |   (_____  )  | |      | |      | |   | |(_____  )|  __)   | |   | |                   \n"
-    "                             | |         ) |  | |      | |      | |   | |      ) || (      | |   ) |                   \n"
-    "                          ___) (___/\\____) |  | (____/\\| (____/\\| (___) |/\\____) || (____/\\| (__/  )                   \n"
-    "                          \\_______/\\_______)  (_______/(_______/(_______)\\_______)(_______/(______/                    \n";
-
-// Animation Masks
-char *CART_HEAD =
-    "             \n"
-    "             \n"
-    "             \n"
-    "      _____  \n"
-    "  II__|[ ]|  \n"
-    " |        |_ \n"
-    "< OO----OOO  \n";
-
-char *CART_BODY =
-    "      o      \n"
-    "    -----    \n"
-    "   /|%03d|\\   \n"
-    "_____________\n"
-    "|           |\n"
-    "|___________|\n"
-    "  OO----OO  \n";
-
-char *CART_EMPTY =
-    "             \n"
-    "             \n"
-    "             \n"
-    "_____________\n"
-    "|  [EMPTY]  |\n"
-    "|___________|\n"
-    "  OO----OO  \n";
-
-char *PERSON_MASK =
-    "   o   \n"
-    " ----- \n"
-    "/|%03d|\\\n"
-    " |   | \n"
-    " ----- \n"
-    "/     \\\n";
-
-char *COASTER_FLOOR =
-    "========================================\n"
-    "|     H*    H   *H         H*    H   *H|\n"
-    "|     H *   H  * H         H *   H  * H|\n"
-    "|     H  *  H *  H         H  *  H *  H|\n"
-    "|     H   * **   H         H   ***    H|\n"
-    "|     H  *  H *  H         H  *  H *  H|\n"
-    "|     H *   H  * H         H *   H  * H|\n"
-    "|     H*    H   *H         H*    H   *H|\n"
-    "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX\n"
-    "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX\n";
-
-char *FLOOR =
-    "##########\n"
-    "oooooooooo\n"
-    "oooooooooo\n"
-    "oooooooooo\n"
-    "oooooooooo\n";
-
-char *ENTRANCE =
-    "---------------\n"
-    "|   ROLLER    |\n"
-    "|   COASTER   |\n"
-    "---------------\n"
-    "||           ||\n"
-    "||           ||\n"
-    "||           ||\n"
-    "||           ||\n"
-    "||           ||\n"
-    "||           ||\n";
-
-char *EXIT =
-    "---------------\n"
-    "|    EXIT     |\n"
-    "|             |\n"
-    "---------------\n"
-    "||           ||\n"
-    "||           ||\n"
-    "||           ||\n"
-    "||           ||\n"
-    "||           ||\n"
-    "||           ||\n";
-
-// Animation Variables
-int CART_HEADX, CART_HEADY;
-PrintBuffer BUF;
-Queue Q;
-Cart CART_LIST[CAPACITY];
 
 // Thread Control Variables
 sem_t mutex;
@@ -170,58 +31,6 @@ sem_t allAboard;
 sem_t allAshore;
 int boarders = 0;
 int unboarders = 0;
-
-// Auxiliary Function
-void add_background(PrintBuffer *buffer)
-{
-    int start_pos = 10;
-    int floor_start_pos = HEIGHT - 5;
-
-    for (int i = 0; i < WIDTH / 40; i++)
-        draw_object(buffer, COASTER_FLOOR, i * 40, start_pos);
-
-    for (int i = 0; i < WIDTH / 10; i++)
-        draw_object(buffer, FLOOR, i * 10, floor_start_pos);
-
-    draw_object(buffer, ENTRANCE, 1, HEIGHT - 15);
-    draw_object(buffer, EXIT, WIDTH - 16, HEIGHT - 15);
-}
-
-void draw_train(PrintBuffer *buffer, Cart *cart_list, int dx)
-{
-    const int space = 12;
-    const int start_pos = 10;
-
-    CART_HEADX += dx;
-    CART_HEADY += 0;
-
-    draw_queue(&Q, buffer);
-    add_background(buffer);
-    draw_object(buffer, CART_HEAD, CART_HEADX, CART_HEADY);
-    for (int i = 0; i < CAPACITY; i++)
-    {
-        cart_list[i].x = cart_list[i].x + dx;
-        draw_cart(&cart_list[i], buffer);
-    }
-}
-
-void move_train(PrintBuffer *buffer, Cart *cart_list)
-{
-    float dx = -WIDTH / TRAIN_MOVE_FPS;
-
-    for (int i = 0; i < WIDTH / abs(dx); i++)
-    {
-        draw_train(buffer, cart_list, dx);
-        render_buffer(buffer, TRAIN_CD);
-    }
-
-    CART_HEADX = CART_STARTX;
-    CART_HEADY = CART_STARTY;
-    for (int i = 0; i < CAPACITY; i++)
-    {
-        set_position(&CART_LIST[i], CART_STARTX + (i + 1) * CART_PADDING, CART_STARTY);
-    }
-}
 
 void load()
 {
