+++ projetomultithread/src/sprites/sprites.c
@@ -1,0 +1,115 @@
+// Title Screens
+char *OPENING =
+    " _______  _______  _        _        _______  _______    _______  _______  _______  _______ _________ _______  _______ \n"
+    "(  ____ )(  ___  )( \\      ( \\      (  ____ \\(  ____ )  (  ____ \\(  ___  )(  ___  )(  ____ \\__   __/(  ____ \\(  ____ )\n"
+    "| (    )|| (   ) || (      | (      | (    \\/| (    )|  | (    \\/| (   ) || (   ) || (    \\/   ) (   | (    \\/| (    )|\n"
+    "| (____)|| |   | || |      | |      | (__    | (____)|  | |      | |   | || (___) || (_____    | |   | (__    | (____)|\n"
+    "|     __)| |   | || |      | |      |  __)   |     __)  | |      | |   | ||  ___  |(_____  )   | |   |  __)   |     __)\n"
+    "| (\\ (   | |   | || |      | |      | (      | (\\ (     | |      | |   | || (   ) |      ) |   | |   | (      | (\\ (   \n"
+    "| ) \\ \\__| (___) || (____/\\| (____/\\| (____/\\| ) \\ \\__  | (____/\\| (___) || )   ( |/\\____) |   | |   | (____/\\| ) \\ \\__\n"
+    "|/   \\__/(_______)(_______/(_______/(_______/|/   \\__/  (_______/(_______)|/     \\|\\_______)   )_(   (_______/|/   \\__/\n"
+    "                             _________ _______    _______  _______  _______  _                                         \n"
+    "                             \\__   __/(  ____ \\  (  ___  )(  ____ )(  ____ \\( (    /|                               \n"
+    "                                ) (   | (    \\/  | (   ) || (    )|| (    \\/|  \\  ( |                               \n"
+    "                                | |   | (_____   | |   | || (____)|| (__    |   \\ | |                                 \n"
+    "                                | |   (_____  )  | |   | ||  _____)|  __)   | (\\ \\) |                                \n"
+    "                                | |         ) |  | |   | || (      | (      | | \\   |                                 \n"
+    "                             ___) (___/\\____) |  | (___) || )      | (____/\\| )  \\  |                               \n"
+    "                             \\_______/\\_______)  (_______)|/       (_______/|/    )_)                                \n";
+
+char *ENDING =
+    " _______  _______  _        _        _______  _______    _______  _______  _______  _______ _________ _______  _______ \n"
+    "(  ____ )(  ___  )( \\      ( \\      (  ____ \\(  ____ )  (  ____ \\(  ___  )(  ___  )(  ____ \\__   __/(  ____ \\(  ____ )\n"
+    "| (    )|| (   ) || (      | (      | (    \\/| (    )|  | (    \\/| (   ) || (   ) || (    \\/   ) (   | (    \\/| (    )|\n"
+    "| (____)|| |   | || |      | |      | (__    | (____)|  | |      | |   | || (___) || (_____    | |   | (__    | (____)|\n"
+    "|     __)| |   | || |      | |      |  __)   |     __)  | |      | |   | ||  ___  |(_____  )   | |   |  __)   |     __)\n"
+    "| (\\ (   | |   | || |      | |      | (      | (\\ (     | |      | |   | || (   ) |      ) |   | |   | (      | (\\ (   \n"
+    "| ) \\ \\__| (___) || (____/\\| (____/\\| (____/\\| ) \\ \\__  | (____/\\| (___) || )   ( |/\\____) |   | |   | (____/\\| ) \\ \\__\n"
+    "|/   \\__/(_______)(_______/(_______/(_______/|/   \\__/  (_______/(_______)|/     \\|\\_______)   )_(   (_______/|/   \\__/\n"
+    "                          _________ _______    _______  _        _______  _______  _______  ______                     \n"
+    "                          \\__   __/(  ____ \\  (  ____ \\( \\      (  ___  )(  ____ \\(  ____ \\(  __  \\                    \n"
+    "                             ) (   | (    \\/  | (    \\/| (      | (   ) || (    \\/| (    \\/| (  \\  )                   \n"
+    "                             | |   | (_____   | |      | |      | |   | || (_____ | (__    | |   ) |                   \n"
+    "                             | |   (_____  )  | |      | |      | |   | |(_____  )|  __)   | |   | |                   \n"
+    "                             | |         ) |  | |      | |      | |   | |      ) || (      | |   ) |                   \n"
+    "                          ___) (___/\\____) |  | (____/\\| (____/\\| (___) |/\\____) || (____/\\| (__/  )                   \n"
+    "                          \\_______/\\_______)  (_______/(_______/(_______)\\_______)(_______/(______/                    \n";
+
+// Animation Masks
+char *CART_HEAD =
+    "             \n"
+    "             \n"
+    "             \n"
+    "      _____  \n"
+    "  II__|[ ]|  \n"
+    " |        |_ \n"
+    "< OO----OOO  \n";
+
+char *CART_BODY =
+    "      o      \n"
+    "    -----    \n"
+    "   /|%03d|\\   \n"
+    "_____________\n"
+    "|           |\n"
+    "|___________|\n"
+    "  OO----OO  \n";
+
+char *CART_EMPTY =
+    "             \n"
+    "             \n"
+    "             \n"
+    "_____________\n"
+    "|  [EMPTY]  |\n"
+    "|___________|\n"
+    "  OO----OO  \n";
+
+char *PERSON_MASK =
+    "   o   \n"
+    " ----- \n"
+    "/|%03d|\\\n"
+    " |   | \n"
+    " ----- \n"
+    "/     \\\n";
+
+char *COASTER_FLOOR =
+    "========================================\n"
+    "|     H*    H   *H         H*    H   *H|\n"
+    "|     H *   H  * H         H *   H  * H|\n"
+    "|     H  *  H *  H         H  *  H *  H|\n"
+    "|     H   * **   H         H   ***    H|\n"
+    "|     H  *  H *  H         H  *  H *  H|\n"
+    "|     H *   H  * H         H *   H  * H|\n"
+    "|     H*    H   *H         H*    H   *H|\n"
+    "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX\n"
+    "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX\n";
+
+char *FLOOR =
+    "##########\n"
+    "oooooooooo\n"
+    "oooooooooo\n"
+    "oooooooooo\n"
+    "oooooooooo\n";
+
+char *ENTRANCE =
+    "---------------\n"
+    "|   ROLLER    |\n"
+    "|   COASTER   |\n"
+    "---------------\n"
+    "||           ||\n"
+    "||           ||\n"
+    "||           ||\n"
+    "||           ||\n"
+    "||           ||\n"
+    "||           ||\n";
+
+char *EXIT =
+    "---------------\n"
+    "|    EXIT     |\n"
+    "|             |\n"
+    "---------------\n"
+    "||           ||\n"
+    "||           ||\n"
+    "||           ||\n"
+    "||           ||\n"
+    "||           ||\n"
+    "||           ||\n";
