+++ projetomultithread/src/sprites/sprites.h
@@ -1,0 +1,26 @@
+#ifndef SPRITES_H
+#define SPRITES_H
+
+// Title Screens
+extern char *OPENING;
+
+extern char *ENDING;
+
+// Animation Masks
+extern char *CART_HEAD;
+
+extern char *CART_BODY;
+
+extern char *CART_EMPTY;
+
+extern char *PERSON_MASK;
+
+extern char *COASTER_FLOOR;
+
+extern char *FLOOR;
+
+extern char *ENTRANCE;
+
+extern char *EXIT;
+    
+#endif
